+++ src/main/java/scanner/XScanRunner.java
@@ -2,11 +2,15 @@ package scanner;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import scanner.scan.CameraScanRangeManager;
+import scanner.http.RangeManager;
 import scanner.scan.CameraScanner;
+import scanner.http.InetSocketAddressRange;
+import scanner.stat.ScanStatGatherer;
 
-import java.net.InetSocketAddress;
 import java.util.List;
+
+import static scanner.stat.ScanStatEnum.ALL;
+import static scanner.stat.ScanStatEnum.RANGES;
 
 /**
  * Scan ip range logic class.
@@ -29,17 +33,20 @@ public class XScanRunner implements Runner {
                 System.out.println("It can be very long. Please, wait...");
                 System.out.println("See log files for more information: /logs/out.log");
 
+                ScanStatGatherer.set(RANGES, listSources.size());
                 for (String range : listSources)
-                    CameraScanRangeManager.prepareSinglePortScanning(range);
-                log.info("addresses will be checked = " + CameraScanRangeManager.count());
+                    RangeManager.prepareSinglePortScanning(range);
+
+                ScanStatGatherer.set(ALL, RangeManager.count());
+                log.info("addresses will be checked = " + RangeManager.count());
 
                 final CameraScanner scanner = new CameraScanner();
-                final List<List<InetSocketAddress>> addressCache = CameraScanRangeManager.getAddressCache();
+                final List<InetSocketAddressRange> addressCache = RangeManager.getAddressCache();
 
                 int c = 0;
-                for (List<InetSocketAddress> listOfIpAddresses : addressCache) {
-                    log.info("progress: ({}) {}/{}", listOfIpAddresses.get(0), ++c, addressCache.size());
-                    scanner.scanning(listOfIpAddresses);
+                for (InetSocketAddressRange range : addressCache) {
+                    log.info("progress: ({}) {}/{}", range.toString(), ++c, addressCache.size());
+                    scanner.scanning(range.list());
                 }
             }
         } catch (Exception xep) {