+++ src/test/java/scanner/CameraScannerTest.java
@@ -2,7 +2,7 @@ package scanner;
 
 import org.junit.Assert;
 import org.junit.Test;
-import scanner.scan.CameraScanRangeManager;
+import scanner.http.RangeManager;
 
 public class CameraScannerTest {
 
@@ -10,8 +10,8 @@ public class CameraScannerTest {
     public void testPrepareSinglePortScanning() {
         String range = "188.202.63.40-188.202.63.47";
 
-        CameraScanRangeManager.prepareSinglePortScanning(range);
-        Assert.assertEquals(8, CameraScanRangeManager.count());
+        RangeManager.prepareSinglePortScanning(range);
+        Assert.assertEquals(8, RangeManager.count());
     }
 
 }